mv modal into void block to avoid key event issues
